Refactor project into multiple namespaces
