+++ RealStereo/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:RealStereo"
-             StartupUri="MainWindow.xaml">
+             StartupUri="Ui/MainWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
