+++ RealStereo/Config/Configuration.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace RealStereo
+namespace RealStereo.Config
 {
     class Configuration
     {
