+++ RealStereo/Config/TestTone.cs
@@ -5,7 +5,7 @@ using NAudio.Wave.SampleProviders;
 using System;
 using System.Collections.Generic;
 
-namespace RealStereo
+namespace RealStereo.Config
 {
     class TestTone
     {
