+++ RealStereo/Ui/ConfigurationWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="RealStereo.ConfigurationWindow"
+﻿<Window x:Class="RealStereo.Ui.ConfigurationWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:RealStereo"
+        xmlns:local="clr-namespace:RealStereo.Ui"
         mc:Ignorable="d"
         Title="Configuration" Height="475" Width="400" Closing="OnClosing">
     <ScrollViewer VerticalScrollBarVisibility="Auto">
