+++ RealStereo/Ui/MainWindow.xaml
@@ -3,8 +3,8 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:RealStereo"
-        xmlns:System="clr-namespace:System;assembly=System.Runtime" x:Class="RealStereo.MainWindow"
+        xmlns:local="clr-namespace:RealStereo.Ui"
+        xmlns:System="clr-namespace:System;assembly=System.Runtime" x:Class="RealStereo.Ui.MainWindow"
         mc:Ignorable="d"
         Title="Real Stereo" Height="520" Width="800" Closed="OnClosed">
     <Grid>
